+++ Private/Test-OpenAIKey.ps1
@@ -1,0 +1,3 @@
+function Test-OpenAIKey {
+    -not [string]::IsNullOrEmpty($env:OpenAIKey)
+}